requirements.txt
Django==5.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
